requirements.txt
django==4.0.1
djangorestframework==3.13.1
markdown==3.3.6
celery==5.2.3
django-celery-results==2.2.0
django-cors-headers==3.11.0
redis==4.1.2
gunicorn==20.1.0
django-heroku==0.3.1
-e ./mailc
